Replace &rpm; entity with RPM
Currently, we don't have much use of &rpm;. It's still
unclear when to use it. As a general concept (in that
case I'd write "RPM") or as a command (we need "rpm").

The entity doesn't provide this distinction so I better
replace them with the appropriate distinction in the text.

Additionally, there is no "RPM-Minimal" image, but a "&bcia;-Minimal"
image.