Fix server order in 2.1 Introduction
Don't know what's exactly meant with "may also need a specific network interface":
do you mean network interface setup?

There are **two** servers required for booting, DHCP is used first:

- PXE boot:  1. DHCP server, 2. TFTP server
- HTTP boot: 1. DHCP server, 2. HTTP server

There are **three** servers required to support both, PXE (TFTP) and HTTP boot.

Further, it is common and recommended to use HTTP to load the agama installer
image (but TFTP or also FTP is supported as well).

Generally, HTTP boot may also use the https instead of the http protocol, but
this requires to apply the certificates in the http server config and also in
the firmware of each client first and is very specific to the client firmware.