remove firewall reload step
The steps above are applying the changes for the current session
**and** are making them permanent (write to config) explicitly,
thus there is no need to reload (load config into current session).

Other possibility would be to use `firewall-cmd --permanent` option
only to write it to config and then a `firewall-cmd --reload` at the
end to (re-)load the new configuration.