remove TFTP file size limitations note
Attic TFTP implementations with 32MB limit aren't usable at all (too small to
load the initrd), 4GB is sufficient for all files (squashfs.img is ~500MB).

Yes, there may be some further limitations of a specific TFTP implementation,
but this is an issue of the (client/firmware) implementation ... and may also
happen to other components.