glibc lifetime: Drop a sentence about documenting dropped symbols
This commit drops a sentence that symbols removed from glibc will be
"properly documented." IMHO it is not useful without describing where
or generally what this means.  The removed symbols will be of course
documented upstream and in release notes which is to be expected, so I
think the sentence is actually mainly superfluous (and may create
expectation that we will document them in some SLES documentation).